move stuff around + stop invalidation from triggering every 10ms
