+++ core/src/cloud/sync/send.rs
@@ -1,0 +1,151 @@
+use super::Library;
+use crate::{cloud::sync::err_break, Node};
+use sd_core_sync::{GetOpsArgs, SyncMessage, NTP64};
+use sd_prisma::prisma::instance;
+use serde::Deserialize;
+use serde_json::json;
+use std::{sync::Arc, time::Duration};
+use tokio::time::sleep;
+use uuid::Uuid;
+
+pub async fn run_actor(library: Arc<Library>, node: Arc<Node>) {
+	let db = &library.db;
+	let api_url = &library.env.api_url;
+	let library_id = library.id;
+
+	loop {
+		println!("send_actor run");
+
+		{
+			// recreate subscription each time so that existing messages are dropped
+			let mut rx = library.sync.subscribe();
+
+			// wait until Created message comes in
+			loop {
+				if let Ok(SyncMessage::Created) = rx.recv().await {
+					break;
+				};
+			}
+		}
+
+		println!("send_actor sleeping");
+
+		sleep(Duration::from_millis(1000)).await;
+
+		println!("send_actor sending");
+
+		loop {
+			let instances = err_break!(
+				db.instance()
+					.find_many(vec![])
+					.select(instance::select!({ pub_id }))
+					.exec()
+					.await
+			)
+			.into_iter()
+			.map(|i| json!({ "instanceUuid": library.instance_uuid.to_string() }))
+			.collect::<Vec<_>>();
+
+			#[derive(Deserialize, Debug)]
+			#[serde(rename_all = "camelCase")]
+			struct RequestAdd {
+				instance_uuid: Uuid,
+				from_time: Option<String>,
+				// mutex key on the instance
+				key: String,
+			}
+
+			let req_adds = err_break!(
+				err_break!(
+					node.authed_api_request(
+						node.http
+							.post(&format!(
+							"{api_url}/api/v1/libraries/{library_id}/messageCollections/requestAdd"
+						))
+							.json(&json!({ "instances": instances })),
+					)
+					.await
+				)
+				.json::<Vec<RequestAdd>>()
+				.await
+			);
+
+			println!("Add Requests: {req_adds:#?}");
+
+			let mut instances = vec![];
+
+			for req_add in req_adds {
+				let ops = err_break!(
+					library
+						.sync
+						.get_ops(GetOpsArgs {
+							count: 50,
+							clocks: vec![(
+								req_add.instance_uuid,
+								NTP64(
+									req_add
+										.from_time
+										.unwrap_or_else(|| "0".to_string())
+										.parse()
+										.unwrap(),
+								),
+							)],
+						})
+						.await
+				);
+
+				if ops.len() == 0 {
+					continue;
+				}
+
+				let start_time = ops[0].timestamp.0.to_string();
+				let end_time = ops[ops.len() - 1].timestamp.0.to_string();
+
+				instances.push(json!({
+					"uuid": req_add.instance_uuid,
+					"key": req_add.key,
+					"startTime": start_time,
+					"endTime": end_time,
+					"contents": ops,
+				}))
+			}
+
+			tracing::debug!("Number of instances: {}", instances.len());
+			tracing::debug!(
+				"Number of messages: {}",
+				instances
+					.iter()
+					.map(|i| i["contents"].as_array().unwrap().len())
+					.sum::<usize>()
+			);
+
+			if instances.len() == 0 {
+				break;
+			}
+
+			#[derive(Deserialize, Debug)]
+			#[serde(rename_all = "camelCase")]
+			struct DoAdd {
+				instance_uuid: Uuid,
+				from_time: String,
+			}
+
+			let responses = err_break!(
+				err_break!(
+					node.authed_api_request(
+						node.http
+							.post(&format!(
+								"{api_url}/api/v1/libraries/{library_id}/messageCollections/doAdd",
+							))
+							.json(&json!({ "instances": instances })),
+					)
+					.await
+				)
+				.json::<Vec<DoAdd>>()
+				.await
+			);
+
+			println!("DoAdd Responses: {responses:#?}");
+		}
+	}
+}
