+++ core/src/library/mod.rs
@@ -1,5 +1,4 @@
 // pub(crate) mod cat;
-mod cloud_sync;
 mod config;
 #[allow(clippy::module_inception)]
 mod library;
